#127 - Add: delete image on ImageSlider component
